implemented custom hook
