+++ client/src/components/CheckoutForm.js
@@ -1,5 +1,9 @@
 import React, { useState } from "react";
 
+//import custom hook
+
+import { useForm } from '../hooks/useForm'
+
 const initialValue = {
   firstName: "",
   lastName: "",
@@ -14,17 +18,10 @@ const initialValue = {
 // and replace the necessary stateful logic from CheckoutForm with the hook
 
 const CheckoutForm = (props) => {
-  const [showSuccessMessage, setShowSuccessMessage] = useState(false);
-  const [values, setValues] = useState(initialValue);
 
-  const handleChanges = (e) => {
-    setValues({ ...values, [e.target.name]: e.target.value });
-  };
+  //using custom useForm hook
+  const [values, handleChanges,handleSubmit, showSuccessMessage ] = useForm(initialValue)
 
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    setShowSuccessMessage(true);
-  };
 
   return (
     <>
