+++ client/src/components/PlantList.js
@@ -14,11 +14,8 @@ export default class PlantList extends Component {
   //   - fetch data from the server endpoint - http://localhost:3333/plants
   //   - set the returned plants array to this.state.plants
   componentDidMount(){
-    console.log('mounted component');
-    
     axios.get('http://localhost:3333/plants')
       .then(res =>{
-        console.log(res.data.plantsData);
         this.setState({
           plants: res.data.plantsData
         })
